implemented review inputs
